+++ config/gh-pages.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# NOTE: This script is a temporary stop-gap measure while we work towards switching
+# to a new site-hosting infrastructure. Ask @lukelooker if you need to deploy the site.
+
+# git co -B gh-pages
+# yarn deploy
+# echo "components.looker.com" > docs/static/CNAME
+# mv docs/static ~/
+# rm -rf *
+# mv ~/static/* .
+# rmdir ~/static # clean-up
+# git add .
+# git commit -a -m "Deploy components.looker.com latest" --no-verify
+# git push origin gh-pages --force