+++ packages/components/src/DateFormat/index.ts
@@ -24,37 +24,4 @@
 
  */
 
-import React, { forwardRef, Ref } from 'react'
-import styled from 'styled-components'
-import { useID } from '../../../utils'
-import { useFormContext } from '../..'
-import { InputDate, InputDateProps } from '../../Inputs/InputDate'
-import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
-
-export interface FieldInputDateProps extends FieldProps, InputDateProps {}
-
-const FieldDateComponent = forwardRef(
-  (props: FieldInputDateProps, ref: Ref<HTMLInputElement>) => {
-    const validationMessage = useFormContext(props)
-    const id = useID(props.id)
-    return (
-      <Field
-        {...pickFieldProps(props)}
-        id={id}
-        validationMessage={validationMessage}
-      >
-        <InputDate
-          {...omitFieldProps(props)}
-          aria-describedby={`${id}-describedby`}
-          id={id}
-          validationType={validationMessage && validationMessage.type}
-          ref={ref}
-        />
-      </Field>
-    )
-  }
-)
-
-FieldDateComponent.displayName = 'FieldDateComponent'
-
-export const FieldDate = styled(FieldDateComponent)``
+export * from './DateFormat'