+++ packages/components/src/DateTimeFormat/DateTimeFormat.test.tsx
@@ -23,22 +23,36 @@
  SOFTWARE.
 
  */
+import { assertSnapshot } from '@looker/components-test-utils'
+import React from 'react'
+import { Locales } from '../utils/i18n'
+import { DateTimeFormat } from './DateTimeFormat'
+const date = new Date('January 25, 1988 11:58:03')
 
-export * from './Checkbox'
-export * from './Combobox'
-export * from './InlineInputText'
-export * from './InlineTextArea'
-export * from './InputChips'
-export * from './InputColor'
-export * from './InputHidden'
-export * from './InputSearch'
-export * from './InputText'
-export * from './InputTime'
-export * from './InputTimeSelect'
-export * from './OptionsGroup'
-export * from './Radio'
-export * from './RangeSlider'
-export * from './Select'
-export * from './Slider'
-export * from './TextArea'
-export * from './ToggleSwitch'
+test('DateTime renders', () => {
+  assertSnapshot(<DateTimeFormat>{date}</DateTimeFormat>)
+})
+
+test('DateTimeFormat renders when passing specific locale', () => {
+  assertSnapshot(
+    <DateTimeFormat locale={Locales.Arabic}>{date}</DateTimeFormat>
+  )
+})
+
+test('DateTimeFormat displays timeZone prop if one is passed', () => {
+  assertSnapshot(
+    <DateTimeFormat timeZone="Asia/Kolkata">{date}</DateTimeFormat>
+  )
+})
+
+test('DateTimeFormat renders format prop short if one is passed', () => {
+  assertSnapshot(<DateTimeFormat format="short">{date}</DateTimeFormat>)
+})
+
+test('DateTimeFormat renders format prop long if one is passed', () => {
+  assertSnapshot(<DateTimeFormat format="long">{date}</DateTimeFormat>)
+})
+
+test('DateTimeFormat renders format prop full if one is passed', () => {
+  assertSnapshot(<DateTimeFormat format="full">{date}</DateTimeFormat>)
+})