+++ packages/icons/src/types.ts
@@ -24,9 +24,7 @@
 
  */
 
-import React, { FC } from 'react'
-import { DateTimeFormatProps, DateTimeFormat } from './DateTimeFormat'
-
-export const TimeFormat: FC<DateTimeFormatProps> = (props) => (
-  <DateTimeFormat {...props} date={false} />
-)
+export interface GlyphProps extends React.SVGProps<SVGSVGElement> {
+  title?: string
+  titleId?: string
+}